+++ FuzzySharp/SimilarityRatio/Scorer/Composite/WeightedRatioScorer.cs
@@ -9,7 +9,7 @@ namespace FuzzySharp.SimilarityRatio.Scorer.Composite
         private static double PARTIAL_SCALE = .90;
         private static bool TRY_PARTIALS = true;
 
-        public override int Score(string input1, string input2)
+        public override double Score(string input1, string input2)
         {
             int len1 = input1.Length;
             int len2 = input2.Length;
@@ -23,7 +23,7 @@ namespace FuzzySharp.SimilarityRatio.Scorer.Composite
             double unbaseScale  = UNBASE_SCALE;
             double partialScale = PARTIAL_SCALE;
 
-            int    baseRatio = Fuzz.Ratio(input1, input2);
+            double baseRatio = Fuzz.Ratio(input1, input2);
             double lenRatio  = ((double) Math.Max(len1, len2)) / Math.Min(len1, len2);
 
             // if strings are similar length don't use partials
@@ -38,14 +38,14 @@ namespace FuzzySharp.SimilarityRatio.Scorer.Composite
                 double partialSor = Fuzz.TokenSortRatio(input1, input2) * unbaseScale * partialScale;
                 double partialSet = Fuzz.TokenSetRatio(input1, input2) * unbaseScale * partialScale;
 
-                return (int) Math.Round(new[] { baseRatio, partial, partialSor, partialSet }.Max());
+                return (double) Math.Round(new[] { baseRatio, partial, partialSor, partialSet }.Max());
             }
             else
             {
                 double tokenSort = Fuzz.TokenSortRatio(input1, input2) * unbaseScale;
                 double tokenSet  = Fuzz.TokenSetRatio(input1, input2) * unbaseScale;
 
-                return (int) Math.Round(new[] { baseRatio, tokenSort, tokenSet }.Max());
+                return (double) Math.Round(new[] { baseRatio, tokenSort, tokenSet }.Max());
             }
         }
