+++ FuzzySharp/SimilarityRatio/Strategy/Generic/DefaultRatioStrategyT.cs
@@ -4,9 +4,9 @@ namespace FuzzySharp.SimilarityRatio.Strategy.Generic
 {
     internal class DefaultRatioStrategy<T> where T : IEquatable<T>
     {
-        public static int Calculate(T[] input1, T[] input2)
+        public static double Calculate(T[] input1, T[] input2)
         {
-            return (int) Math.Round(100 * Levenshtein.GetRatio(input1, input2));
+            return (double) Math.Round(100 * Levenshtein.GetRatio(input1, input2));
         }
     }
 }
